pyproject.toml
[tool.poetry]
name = "paper2ppt"
version = "0.1.0"
description = ""
authors = ["Abhay Singhal <20944075+sAbhay@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
tabula-py = "^2.8.2"
pymupdf = "^1.23.6"
arxiv = "^2.0.0"
pylatexenc = "^2.10"
pillow = "^10.1.0"
pdf2image = "^1.16.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

